rename: lottomodel 파일명 변경
